rename 'communities' session
