+++ src/App.tsx
@@ -3,7 +3,7 @@ import Header from './components/header/header'
 import About from './components/about/about'
 import Participate from './components/participate/participate'
 import Speakers from './components/speakers/speakers'
-import Sponsors from './components/sponsors/sponsors'
+import Community from './components/community/community'
 import Footer from './components/footer/footer'
 
 function App() {
@@ -13,7 +13,7 @@ function App() {
       <About />
       <Participate />
       <Speakers />
-      <Sponsors />
+      <Community />
       <Footer />
     </>
   )
