+++ src/components/community/community.scss
